paraMed_pipeline/requirements.txt
requests
beautifulsoup4
pymongo
sentence-transformers
scikit-learn
python-dotenv